requirements.txt
Flask==2.2.2
python-telegram-bot==20.0
schedule==1.1.0
selenium==4.8.0
requests==2.28.1